requirements.txt
FastAPI==0.95.1
uvicorn==0.22.0
selenium==4.9.0
requests==2.28.2
openai==0.28.1
langchain==0.0.330
langsmith==0.0.57
python-dotenv==1.0.0
pushbullet.py==0.12.0
schedule==1.2.0
tenacity==8.2.3